Change ash filesystem WriteFile operation to take size_t length
A future CL will enable checks that all sizes passed to
base::span are of type size_t.

ash filesystem provider for WriteFile would previously take
a signed `length` value and pass this into base::make_span
which could potentially be signed and thus invalid.

This CL changes this behavior so that the WriteFile operation
takes a size_t, and the only caller checks that the length
is valid before creating the operation object.

In the event of an invalid length, a FILE_ERROR_INVALID_OPERATION
is returned to the caller.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4216195
Commit-Queue: Will Harris <wfh@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1101875}

+++ chrome/browser/ash/file_system_provider/operations/write_file.h
@@ -22,7 +22,7 @@ namespace ash {
 namespace file_system_provider {
 namespace operations {
 
-// Bridge between fileapi write file and providing extension's write fil
+// Bridge between fileapi write file and providing extension's write file
 // request.
 // Created per request.
 class WriteFile : public Operation {
@@ -32,7 +32,7 @@ class WriteFile : public Operation {
             int file_handle,
             scoped_refptr<net::IOBuffer> buffer,
             int64_t offset,
-            int length,
+            size_t length,
             storage::AsyncFileUtil::StatusCallback callback);
 
   WriteFile(const WriteFile&) = delete;
@@ -53,7 +53,7 @@ class WriteFile : public Operation {
   int file_handle_;
   scoped_refptr<net::IOBuffer> buffer_;
   int64_t offset_;
-  int length_;
+  size_t length_;
   storage::AsyncFileUtil::StatusCallback callback_;
 };
